Change tag's type of room messages. Now, it's a byte

+++ sdk/src/com/scoreflex/realtime/Message.java
@@ -102,8 +102,8 @@ public class Message {
    *
    * @return The message's tag.
    */
-  public int getTag() {
-    return tag;
+  public byte getTag() {
+    return (byte)tag;
   }
 
   /**
